General Updates - Testing

+++ NWS.net/Main.cs
@@ -4,13 +4,13 @@ class Program {
 
     public static NWS.net.NWSAPI API;
 
-    static float Lat;
-    static float Lon;
+    static double Lat;
+    static double Lon;
 
     static void Main(string[] args) {
         if(args.Length != 3) { Console.WriteLine("You must enter 3 arguments"); return; }
-        Lat = float.Parse(args[0]);
-        Lon = float.Parse(args[1]);
+        Lat = double.Parse(args[0]);
+        Lon = double.Parse(args[1]);
         int Update = int.Parse(args[2]);
 
         System.Timers.Timer T = new System.Timers.Timer(1000 * Update);
